Remove outdated code and functions
