+++ dkg/managers/async_manager.py
@@ -19,8 +19,8 @@ from typing import Any, Type
 
 from dkg.dataclasses import BlockchainResponseDict, NodeResponseDict
 from dkg.exceptions import InvalidRequest
-from dkg.utils.blockchain_request import ContractInteraction, JSONRPCRequest
-from dkg.utils.node_request import NodeCall
+from dkg.request_managers.blockchain_request import ContractInteraction, JSONRPCRequest
+from dkg.request_managers.node_request import NodeCall
 from dkg.providers.node.async_node_http import AsyncNodeHTTPProvider
 from dkg.providers.blockchain.async_blockchain import AsyncBlockchainProvider
