--- dkg/utils/knowledge_asset_tools.py
@@ -1,5 +1,0 @@
-from uuid import uuid4
-
-
-def generate_named_node():
-    return f"uuid:{uuid4()}"
